refactor(api): split chatGraphContractRouter into context + resume modules
Slim the 1352-line contract router:
- services/streamContext.ts: buildStreamContext() owns all pre-classify setup
  (auth/validation, notebook resolution, message conversion, thread creation,
  attachment processing, memory, state init, context enrichment) and returns a
  discriminated { done } | { done:false, ctx } result.
- services/resumePipeline.ts: runChatGraphResume() holds the full resume flow.
- sseHelpers.ts: new sseFail() consolidates the send-error/end/return pattern.

Router keeps Stages 1-4 inline (reading from the destructured ctx) and becomes
thin wiring. Handler arg/body types derived via ts-rest ServerInferRequest.
No behavior change.
