+++ apps/api/routes/chat/chatGraphContractRouter.ts
@@ -10,51 +10,27 @@
  * provides typed request-body validation; the SSE stream itself is opaque
  * from ts-rest's perspective.
  *
+ * The handlers are kept thin:
+ * - `stream` builds the request context (./services/streamContext) and then
+ *   runs Stages 1–4 (classify → intent → response → persist) inline.
+ * - `resume` delegates wholesale to ./services/resumePipeline.
+ *
  * Mount this BEFORE the legacy router in routes.ts so ts-rest matches
  * its own routes first; unmatched paths fall through to the legacy router.
  */
 
 import { chatGraphContract } from '@gruenerator/contracts';
 import { createExpressEndpoints, initServer } from '@ts-rest/express';
-import { convertToModelMessages } from 'ai';
 
-import {
-  initializeChatState,
-  buildSystemMessage,
-  classifierNode,
-  briefGeneratorNode,
-  searchNode,
-  rerankNode,
-  buildCitations,
-} from '../../agents/langgraph/ChatGraph/index.js';
-import {
-  isKnownNotebook,
-  isUserNotebookId,
-  resolveUserNotebookDocumentIds,
-} from '../../config/notebookCollectionMap.js';
+import { classifierNode, buildSystemMessage } from '../../agents/langgraph/ChatGraph/index.js';
 import { isRegoloReasoningModel } from '../../services/ai/regoloReasoningStream.js';
-import {
-  getMem0Instance,
-  normalizeCategory,
-  formatMemoriesByCategory,
-} from '../../services/mem0/index.js';
-import { getCachedPersona } from '../../services/mem0/personaService.js';
 import { logContractValidationError } from '../../utils/contractValidationLogger.js';
-import { getAIWorkerPool } from '../../utils/getAIWorkerPool.js';
-import { NextcloudShareManager } from '../../utils/integrations/nextcloud/shareManager.js';
 import { createLogger } from '../../utils/logger.js';
-import { ThreadId, UserId } from '../../utils/types/branded.js';
 
-import { getContextWindow } from './agents/providers.js';
-import { getThreadAttachments } from './services/attachmentPersistenceService.js';
-import {
-  processAttachments,
-  injectImageAttachments,
-} from './services/attachmentProcessingService.js';
+import { injectImageAttachments } from './services/attachmentProcessingService.js';
 import { searchChatHistory } from './services/chatSearchService.js';
 import { extractCompoundTopic } from './services/compoundTopicExtractor.js';
 import { extractChartFromResponse, emitConfirmAction } from './services/confirmActionService.js';
-import { enrichContext } from './services/contextEnrichmentService.js';
 import { pruneMessages, applyCompaction } from './services/contextPruningService.js';
 import {
   handleBoardCreation,
@@ -62,18 +38,16 @@ import {
   handleShareDoc,
   executeIntentPipeline,
 } from './services/intentExecutionService.js';
-import { extractTextContent, filterEmptyAssistantMessages } from './services/messageHelpers.js';
+import { extractTextContent } from './services/messageHelpers.js';
 import { pipelineStateStore } from './services/pipelineStateStore.js';
-import {
-  persistAssistantResponse,
-  persistResumedResponse,
-} from './services/postResponseService.js';
+import { persistAssistantResponse } from './services/postResponseService.js';
 import {
   resolveModel,
   buildMessagesForAI,
   streamForResolution,
   streamWithFallback,
 } from './services/responseStreamingService.js';
+import { runChatGraphResume } from './services/resumePipeline.js';
 import {
   getLastSharepicVariant,
   isSharepicRefinement,
@@ -81,15 +55,10 @@ import {
   type PriorSharepic,
 } from './services/sharepicVariantHelpers.js';
 import { createSSEStream, getIntentMessage, PROGRESS_MESSAGES } from './services/sseHelpers.js';
-import { canAccessThread } from './services/threadAccessService.js';
-import { getUser, createThread, createMessage } from './services/threadPersistenceService.js';
+import { buildStreamContext } from './services/streamContext.js';
 
-import type {
-  ChatGraphState,
-  ChatGraphInput,
-  ProcessedAttachment,
-} from '../../agents/langgraph/ChatGraph/types.js';
-import type { ModelMessage, UIMessage } from 'ai';
+import type { ChatGraphState } from '../../agents/langgraph/ChatGraph/types.js';
+import type { ModelMessage } from 'ai';
 import type { Application } from 'express';
 
 const log = createLogger('chatGraphContractRouter');
@@ -98,371 +67,47 @@ const s = initServer();
 
 export const chatGraphContractRouter = s.router(chatGraphContract, {
   stream: async (args) => {
-    const { req, res } = args;
-    const sse = createSSEStream(res);
+    const { req } = args;
+    const sse = createSSEStream(args.res);
     const requestId = `req_${Date.now()}`;
     log.info('[chatGraphContract] stream handler entered, request_id=%s', requestId);
 
     try {
+      const ctxResult = await buildStreamContext({ req, body: args.body, sse, requestId });
+      if (ctxResult.done) {
+        return { status: 200 as const, body: undefined };
+      }
+
+      // Destructure the context into the same identifiers the staged pipeline
+      // below was written against (requestId is already in scope from above).
       const {
-        messages: clientMessages,
-        agentId,
-        threadId,
-        enabledTools,
-        modelId,
-        attachments,
-        notebookIds: rawNotebookIds,
-        forcedTools,
-        documentIds: rawDocumentIds,
-        textIds: rawTextIds,
-        documentChatIds: rawDocumentChatIds,
-        documentChatMode,
-        attachmentContext: rawClientAttachmentContext,
-        defaultNotebookId: rawDefaultNotebookId,
-        boardIds: rawBoardIds,
-        docMentionIds: rawDocMentionIds,
-        wolkeFiles: rawWolkeFiles,
-        connectFiles: rawConnectFiles,
-        currentDocument: rawCurrentDocument,
-        customSystemPrompt: rawCustomSystemPrompt,
-        roleName: rawRoleName,
-        initialAssistantMessage: rawInitialAssistantMessage,
-        activeSkillMention: rawActiveSkillMention,
-      } = args.body;
-
-      // === Validate ===
-      const user = getUser(req);
-      if (!user?.id) {
-        sse.send('error', { error: PROGRESS_MESSAGES.unauthorized });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const userId = user.id;
-      const aiWorkerPool = getAIWorkerPool(req);
-
-      if (!aiWorkerPool) {
-        sse.send('error', { error: PROGRESS_MESSAGES.aiUnavailable });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      if ((clientMessages as unknown[]).length === 0) {
-        sse.send('error', { error: PROGRESS_MESSAGES.messagesRequired });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const systemNotebookIds = rawNotebookIds?.filter(isKnownNotebook) ?? [];
-      const userNotebookUuids = rawNotebookIds?.filter(isUserNotebookId) ?? [];
-      const { documentIds: notebookDocumentIds, resolvedUserNotebookIds } =
-        userNotebookUuids.length > 0
-          ? await resolveUserNotebookDocumentIds(userId, userNotebookUuids)
-          : { documentIds: [], resolvedUserNotebookIds: [] };
-      const notebookIds = [...systemNotebookIds, ...resolvedUserNotebookIds];
-      const defaultNotebookId =
-        rawDefaultNotebookId && isKnownNotebook(rawDefaultNotebookId)
-          ? rawDefaultNotebookId
-          : undefined;
-      // An agent can bind a user-owned notebook (UUID) as its default knowledge
-      // base. Resolve it to document IDs (ownership-checked) so search can scope
-      // to it — mirrors the mention path, but as a default rather than explicit.
-      const { documentIds: defaultNotebookDocumentIds } =
-        rawDefaultNotebookId && !defaultNotebookId && isUserNotebookId(rawDefaultNotebookId)
-          ? await resolveUserNotebookDocumentIds(userId, [rawDefaultNotebookId])
-          : { documentIds: [] };
-
-      // Filter wolkeFiles to refs whose shareLinkId is still owned + active for this user.
-      // Stale refs (deleted/deactivated share link) are dropped silently so the chat still works.
-      let wolkeFiles: typeof rawWolkeFiles = undefined;
-      if (rawWolkeFiles?.length) {
-        try {
-          const userShareLinks = await NextcloudShareManager.getShareLinks(userId);
-          const allowedIds = new Set(userShareLinks.filter((l) => l.is_active).map((l) => l.id));
-          const filtered = rawWolkeFiles.filter((f) => allowedIds.has(f.shareLinkId));
-          wolkeFiles = filtered.length > 0 ? filtered : undefined;
-          if (filtered.length < rawWolkeFiles.length) {
-            log.warn(
-              `[ChatGraph] Dropped ${rawWolkeFiles.length - filtered.length} stale wolkeFiles ref(s) for user ${userId}`
-            );
-          }
-        } catch (err) {
-          log.warn(`[ChatGraph] wolkeFiles ownership check failed; ignoring refs`, err);
-          wolkeFiles = undefined;
-        }
-      }
-
-      // @connect file refs need no per-ref ownership pre-check: the Nango
-      // connection (resolved per-file at retrieval time via
-      // ConnectionService.getConnection(userId, provider)) IS the ownership
-      // boundary, and a revoked/expired token fails safe to an empty result.
-      // Normalize null mimeType → omit so downstream types stay clean.
-      const connectFiles = rawConnectFiles?.length
-        ? rawConnectFiles.map((f) => ({
-            provider: f.provider,
-            fileId: f.fileId,
-            name: f.name,
-            ...(f.mimeType ? { mimeType: f.mimeType } : {}),
-          }))
-        : undefined;
-
-      log.info(`[ChatGraph] Processing request for user ${userId}, agent ${agentId ?? 'default'}`);
-      if (notebookIds.length > 0) {
-        log.info(`[ChatGraph] Notebook scoping: ${notebookIds.join(', ')}`);
-      }
-      if (resolvedUserNotebookIds.length > 0) {
-        log.info(
-          `[ChatGraph] User-notebook scoping: ${resolvedUserNotebookIds.length} notebook(s) → ${notebookDocumentIds.length} document(s)`
-        );
-      }
-
-      // === Convert messages ===
-      let modelMessages: ChatGraphInput['messages'];
-      try {
-        modelMessages = (await convertToModelMessages(
-          clientMessages as UIMessage[]
-        )) as ModelMessage[] as ChatGraphInput['messages'];
-      } catch (convertError) {
-        log.error('[ChatGraph] Error converting messages:', convertError);
-        sse.send('error', { error: 'Failed to process messages' });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      if (!modelMessages || !Array.isArray(modelMessages)) {
-        sse.send('error', { error: 'Failed to process messages' });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const validMessages = filterEmptyAssistantMessages(
-        modelMessages as ModelMessage[]
-      ) as ChatGraphInput['messages'];
-      log.info(
-        `[ChatGraph] Converted ${(clientMessages as unknown[]).length} → ${validMessages.length} valid messages`
-      );
-
-      const lastUserMessage = validMessages.filter((m) => m.role === 'user').pop();
-
-      // === Create thread if needed ===
-      // Normalize null → undefined: contract schema uses .nullish() to accept
-      // both, but downstream code is typed for string | undefined.
-      let actualThreadId: string | undefined = threadId ?? undefined;
-      let isNewThread = false;
-
-      if (!actualThreadId && lastUserMessage) {
-        const userText = extractTextContent(lastUserMessage.content);
-        const thread = await createThread(
-          userId,
-          agentId ?? 'gruenerator-universal',
-          userText.slice(0, 50) + (userText.length > 50 ? '...' : '') || 'Neue Unterhaltung'
-        );
-        actualThreadId = thread.id;
-        isNewThread = true;
-        sse.send('thread_created', { threadId: actualThreadId });
-      }
-
-      if (actualThreadId && lastUserMessage) {
-        if (!isNewThread) {
-          if (!(await canAccessThread(ThreadId(actualThreadId), UserId(userId)))) {
-            sse.send('error', { error: 'Thread not found' });
-            sse.end();
-            return { status: 200 as const, body: undefined };
-          }
-        }
-
-        // Seed message (Antrag / PM / Social text) — persisted BEFORE the user
-        // message so order is seed → user → assistant-reply. New threads only.
-        if (
-          isNewThread &&
-          typeof rawInitialAssistantMessage === 'string' &&
-          rawInitialAssistantMessage
-        ) {
-          await createMessage(
-            actualThreadId,
-            'assistant',
-            rawInitialAssistantMessage,
-            { seed: true },
-            userId
-          );
-        }
-
-        const userText = extractTextContent(lastUserMessage.content);
-        await createMessage(
-          actualThreadId,
-          'user',
-          userText,
-          rawRoleName ? { roleName: rawRoleName } : undefined,
-          userId
-        );
-      }
-
-      // Light progress ping so the UI shows "Verstehe Anfrage…" immediately,
-      // instead of looking frozen during attachment processing + memory fetch
-      // + classification (collectively 1–8s on a cold path).
-      const classifyStepId = `classify_${Date.now()}`;
-      sse.send('progress_step', {
-        stepId: classifyStepId,
-        toolName: 'classify',
-        title: 'Verstehe Anfrage…',
-        status: 'in_progress',
-      });
-
-      // === Process attachments ===
-      const {
-        attachmentContext: derivedAttachmentContext,
+        userId,
+        aiWorkerPool,
+        notebookIds,
+        validMessages,
+        lastUserMessage,
+        actualThreadId,
+        isNewThread,
+        classifyStepId,
         imageAttachments,
         processedMeta,
-      } = await processAttachments(attachments as ProcessedAttachment[] | undefined, requestId);
-
-      // Merge any client-injected context (e.g. docs editor markdown + selection)
-      // with what processAttachments derived from uploaded files.
-      const clientAttachmentContext =
-        (rawClientAttachmentContext as string | null | undefined)?.trim() || undefined;
-      const attachmentContext =
-        clientAttachmentContext && derivedAttachmentContext
-          ? `${clientAttachmentContext}\n\n---\n\n${derivedAttachmentContext}`
-          : clientAttachmentContext || derivedAttachmentContext;
-
-      const docAttachments =
-        (attachments as ProcessedAttachment[] | undefined)?.filter((a) => !a.isImage) ?? [];
-
-      const previousAttachments = actualThreadId
-        ? await getThreadAttachments(actualThreadId, 5)
-        : [];
-
-      // === Memory retrieval (mem0) ===
-      let memoryContext: string | null = null;
-      let memoryRetrieveTimeMs = 0;
-      let memoriesUsed: Array<{ content: string; category: string | null }> = [];
-
-      const mem0 = getMem0Instance();
-      if (mem0 && lastUserMessage) {
-        try {
-          const memoryStartTime = Date.now();
-
-          const persona = await getCachedPersona(userId);
-          if (persona) {
-            memoryContext = persona;
-            memoriesUsed = [{ content: '[Persona]', category: null }];
-            log.info(`[${requestId}] Using cached persona for memory context`);
-          } else {
-            const userQuery = extractTextContent(lastUserMessage.content);
-            const memories = await mem0.searchMemories(userQuery, userId, 5);
-            if (memories.length > 0) {
-              memoriesUsed = memories.map((m) => ({
-                content: m.memory,
-                category: normalizeCategory(m.metadata?.memoryType) ?? null,
-              }));
-
-              memoryContext = formatMemoriesByCategory(
-                memories.map((m) => ({
-                  memory: m.memory,
-                  category: normalizeCategory(m.metadata?.memoryType),
-                }))
-              );
-              log.info(`[${requestId}] Retrieved ${memories.length} memories for context`);
-            }
-          }
-
-          memoryRetrieveTimeMs = Date.now() - memoryStartTime;
-        } catch (memError) {
-          log.warn(`[${requestId}] Memory retrieval failed (continuing without):`, memError);
-        }
-      }
-
-      // === Read user profile instructions ===
-      const userInstructions = user.custom_prompt?.trim() || undefined;
-
-      // === Resolve context window for model-aware budgets ===
-      const contextWindowTokens = getContextWindow(modelId);
-
-      // === Initialize state ===
-      const initialState = await initializeChatState({
-        messages: validMessages,
-        threadId: actualThreadId,
-        agentId: agentId ?? 'gruenerator-universal',
-        userId,
-        enabledTools: enabledTools ?? {
-          search: true,
-          web: true,
-          person: true,
-          examples: true,
-          research: true,
-          image: true,
-          image_edit: true,
-        },
-        aiWorkerPool,
-        attachmentContext: attachmentContext ?? undefined,
-        imageAttachments: imageAttachments.length > 0 ? imageAttachments : undefined,
-        threadAttachments: previousAttachments.length > 0 ? previousAttachments : undefined,
-        notebookIds: notebookIds.length > 0 ? notebookIds : undefined,
-        notebookDocumentIds: notebookDocumentIds.length > 0 ? notebookDocumentIds : undefined,
-        defaultNotebookId,
-        defaultNotebookDocumentIds:
-          defaultNotebookDocumentIds.length > 0 ? defaultNotebookDocumentIds : undefined,
-        documentIds: rawDocumentIds?.length ? rawDocumentIds : undefined,
-        documentChatIds: rawDocumentChatIds?.length
-          ? rawDocumentChatIds
-          : docAttachments.length > 0 || documentChatMode
-            ? []
-            : undefined,
-        boardIds: rawBoardIds?.length ? rawBoardIds : undefined,
-        wolkeFiles,
-        connectFiles,
-        // When the docs editor sends a currentDocument, also surface its id as a
-        // doc-mention so the existing modify_doc / summary intent paths activate
-        // (they key off `hasDocMentions`). Explicit @doc mentions always take
-        // precedence — we only fall back to currentDocument.id otherwise.
-        docMentionIds: rawDocMentionIds?.length
-          ? rawDocMentionIds
-          : rawCurrentDocument?.id
-            ? [rawCurrentDocument.id]
-            : undefined,
-        currentDocument: rawCurrentDocument
-          ? {
-              id: rawCurrentDocument.id,
-              title: rawCurrentDocument.title ?? null,
-              markdown: rawCurrentDocument.markdown,
-              selectionText: rawCurrentDocument.selectionText ?? null,
-            }
-          : undefined,
-        userLocale: user.locale ?? 'de-DE',
-        customSystemPrompt: rawCustomSystemPrompt ?? undefined,
-        activeSkillMention: rawActiveSkillMention ?? undefined,
-        userInstructions,
-        contextWindowTokens,
-      });
-
-      const userLocale = user.locale ?? 'de-DE';
-      log.info(`[ChatGraph] User ${userId} locale: ${userLocale}`);
-
-      initialState.agentConfig.userId = userId;
-      if (memoryContext) {
-        initialState.memoryContext = memoryContext;
-        initialState.memoryRetrieveTimeMs = memoryRetrieveTimeMs;
-
-        const isPersona = memoriesUsed.length === 1 && memoriesUsed[0].content === '[Persona]';
-        sse.send('memory_context', {
-          memoryCount: isPersona ? 1 : memoriesUsed.length,
-          memories: isPersona ? [] : memoriesUsed,
-          isPersona,
-        });
-      }
-
-      // === Enrich context (documents, boards, mentions, vectorization) ===
-      await enrichContext({
         initialState,
-        userId,
-        ...(rawDocumentIds != null && { rawDocumentIds }),
-        ...(rawTextIds != null && { rawTextIds }),
-        ...(rawBoardIds != null && { rawBoardIds }),
-        ...(rawDocMentionIds != null && { rawDocMentionIds }),
-        docAttachments,
-        processedMeta,
+        memoryContext,
+        memoryRetrieveTimeMs,
         contextWindowTokens,
-        sse,
-      });
+      } = ctxResult.ctx;
+
+      const {
+        agentId,
+        forcedTools,
+        enabledTools,
+        modelId,
+        documentIds: rawDocumentIds,
+        documentChatIds: rawDocumentChatIds,
+        docMentionIds: rawDocMentionIds,
+        boardIds: rawBoardIds,
+        currentDocument: rawCurrentDocument,
+      } = args.body;
 
       // === Stage 1: Classify ===
       const classifiedState = {
@@ -1074,270 +719,8 @@ export const chatGraphContractRouter = s.router(chatGraphContract, {
   },
 
   resume: async (args) => {
-    const { req } = args;
     const sse = createSSEStream(args.res);
-    const _requestId = `resume_${Date.now()}`;
-    log.info('[chatGraphContract] resume handler entered, request_id=%s', _requestId);
-
-    try {
-      const { threadId, resume: userAnswer } = args.body;
-
-      const user = getUser(req);
-      if (!user?.id) {
-        sse.send('error', { error: PROGRESS_MESSAGES.unauthorized });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const stored = await pipelineStateStore.get(threadId);
-      if (!stored) {
-        sse.send('error', {
-          error: 'Pipeline-Status abgelaufen. Bitte sende deine Nachricht erneut.',
-        });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-      await pipelineStateStore.delete(threadId);
-
-      const { classifiedState, requestContext } = stored;
-
-      if (requestContext.userId !== user.id) {
-        sse.send('error', { error: PROGRESS_MESSAGES.unauthorized });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const aiWorkerPool = getAIWorkerPool(req);
-      if (!aiWorkerPool) {
-        sse.send('error', { error: PROGRESS_MESSAGES.aiUnavailable });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      log.info(`[ChatGraph:Resume] Thread ${threadId}, answer: "${userAnswer.slice(0, 80)}"`);
-
-      classifiedState.needsClarification = false;
-      classifiedState.searchQuery = userAnswer;
-      classifiedState.clarificationQuestion = null;
-      classifiedState.clarificationOptions = null;
-
-      const startTime = Date.now();
-
-      // === Sharepic resume: the answer is the topic — regenerate and finish ===
-      if (classifiedState.intent === 'sharepic') {
-        // Combine the original (topic-less) request with the answer so any variant
-        // hint ("zitat sharepic") survives and the answer supplies the subject.
-        const prevUserMsg = [...classifiedState.messages].reverse().find((m) => m.role === 'user');
-        const prevText = prevUserMsg ? extractTextContent(prevUserMsg.content) : '';
-        const combined = `${prevText} ${userAnswer}`.trim();
-        classifiedState.messages = [
-          ...classifiedState.messages,
-          { role: 'user', content: combined },
-        ];
-
-        sse.send('intent', {
-          intent: 'sharepic',
-          message: getIntentMessage('sharepic'),
-          reasoning: `Resumed: ${userAnswer}`,
-        });
-
-        const { sharepicVariants } = await executeIntentPipeline({
-          classifiedState,
-          sse,
-          forcedTool: requestContext.forcedTool,
-          ...(requestContext.enabledTools != null && { enabledTools: requestContext.enabledTools }),
-          imageAttachments: requestContext.imageAttachments ?? [],
-          req,
-        });
-
-        const n = sharepicVariants.length;
-        const fullText =
-          n > 0
-            ? `Ich habe dir ${n} Sharepic-${n === 1 ? 'Variante' : 'Varianten'} erstellt. ` +
-              `Wähle eine aus oder sag mir, was ich am Text oder Bild anpassen soll.`
-            : `Die Sharepic-Erstellung hat leider nicht geklappt. Magst du es mit einem ` +
-              `anderen Thema noch einmal versuchen?`;
-        sse.send('response_start', { message: PROGRESS_MESSAGES.responseStart });
-        sse.send('text_delta', { text: fullText });
-
-        await persistResumedResponse({
-          threadId: requestContext.actualThreadId!,
-          fullText,
-          finalState: classifiedState,
-          classifiedState,
-        });
-
-        sse.send('done', {
-          ...(requestContext.actualThreadId != null && {
-            threadId: requestContext.actualThreadId,
-          }),
-          citations: [],
-          metadata: {
-            intent: 'sharepic',
-            searchCount: 0,
-            totalTimeMs: Date.now() - startTime,
-            searchTimeMs: 0,
-          },
-        });
-        sse.end();
-        return { status: 200 as const, body: undefined };
-      }
-
-      const nonSearchIntents = new Set(['direct', 'image', 'image_edit']);
-      if (nonSearchIntents.has(classifiedState.intent)) {
-        classifiedState.intent = 'search';
-      }
-
-      sse.send('intent', {
-        intent: classifiedState.intent,
-        message: getIntentMessage(classifiedState.intent),
-        reasoning: `Resumed: ${userAnswer}`,
-        ...(classifiedState.searchQuery != null && { searchQuery: classifiedState.searchQuery }),
-        ...(classifiedState.subQueries != null && { subQueries: classifiedState.subQueries }),
-        ...(classifiedState.searchSources?.length && {
-          searchSources: classifiedState.searchSources,
-        }),
-      });
-
-      // === Search ===
-      let finalState = classifiedState;
-      const { enabledTools, modelId, forcedTool } = requestContext;
-
-      if (!nonSearchIntents.has(classifiedState.intent)) {
-        const toolEnabled = forcedTool || enabledTools?.[classifiedState.intent] !== false;
-        if (toolEnabled) {
-          let searchInputState = classifiedState;
-          if (classifiedState.complexity === 'complex' && classifiedState.intent === 'research') {
-            const briefResult = await briefGeneratorNode(classifiedState);
-            searchInputState = { ...classifiedState, ...briefResult } as ChatGraphState;
-          }
-
-          sse.send('search_start', { message: PROGRESS_MESSAGES.searchStart });
-          const searchResult = await searchNode(searchInputState);
-          finalState = { ...searchInputState, ...searchResult } as ChatGraphState;
-
-          if (finalState.searchResults?.length > 3) {
-            const rerankResult = await rerankNode(finalState);
-            finalState = { ...finalState, ...rerankResult } as ChatGraphState;
-            if (finalState.searchResults.length > 0) {
-              finalState.citations = buildCitations(finalState.searchResults);
-            }
-          }
-
-          const resultCount = finalState.searchResults?.length || 0;
-          sse.send('search_complete', {
-            message: PROGRESS_MESSAGES.searchComplete(resultCount),
-            resultCount,
-            results:
-              finalState.searchResults?.slice(0, 10).map((r) => {
-                const result: {
-                  source: string;
-                  title: string;
-                  content: string;
-                  url?: string;
-                  relevance?: number;
-                } = {
-                  source: r.source,
-                  title: r.title,
-                  content: r.content,
-                };
-                if (r.url != null) result.url = r.url;
-                if (r.relevance != null) result.relevance = r.relevance;
-                return result;
-              }) ?? [],
-            ...((classifiedState.intent === 'examples' ||
-              classifiedState.intent === 'pressemitteilung_examples') &&
-            finalState.examplesResult
-              ? { examplesResult: finalState.examplesResult }
-              : {}),
-          });
-        }
-      }
-
-      // === Response ===
-      sse.send('response_start', { message: PROGRESS_MESSAGES.responseStart });
-
-      const systemMessage = await buildSystemMessage(finalState);
-      const resumeImageAttachments = requestContext.imageAttachments ?? [];
-      const agentConfigForResolve2 = {
-        provider: finalState.agentConfig.provider as string,
-        model: finalState.agentConfig.model,
-        ...(finalState.agentConfig.defaultModel != null && {
-          defaultModel: finalState.agentConfig.defaultModel,
-        }),
-      };
-      const resumeRequestId = `resume_contract_${Date.now()}`;
-      const resolution2 = await resolveModel(agentConfigForResolve2, modelId, resumeRequestId, {
-        hasImages: resumeImageAttachments.length > 0,
-        intent: finalState.intent,
-      });
-      const validMessages = requestContext.validMessages;
-      const prunedValidMessages = pruneMessages(validMessages);
-      const messagesForAI = buildMessagesForAI(systemMessage, prunedValidMessages);
-
-      const baseMaxTokens = finalState.agentConfig.params.max_tokens;
-
-      let fullText: string | null;
-      try {
-        fullText = await streamWithFallback({
-          primary: resolution2,
-          sse,
-          logPrefix: '[ChatGraph:Resume]',
-          buildStream: async (r) => {
-            const isReasoning = isRegoloReasoningModel(r.provider, r.modelName);
-            return streamForResolution({
-              resolution: r,
-              messages: messagesForAI,
-              maxTokens: isReasoning ? Math.max(baseMaxTokens, 9000) : baseMaxTokens,
-              temperature: finalState.agentConfig.params.temperature,
-              sse,
-              logPrefix: '[ChatGraph:Resume]',
-            });
-          },
-        });
-      } finally {
-        if (resolution2.releaseSlot) await resolution2.releaseSlot();
-      }
-
-      if (fullText === null) return { status: 200 as const, body: undefined };
-
-      // === Persist & complete ===
-      await persistResumedResponse({
-        threadId: requestContext.actualThreadId!,
-        fullText,
-        finalState,
-        classifiedState,
-      });
-
-      const totalTimeMs = Date.now() - startTime;
-      sse.send('done', {
-        ...(requestContext.actualThreadId != null && { threadId: requestContext.actualThreadId }),
-        citations: finalState.citations,
-        metadata: {
-          intent: finalState.intent,
-          searchCount: finalState.searchCount || 0,
-          totalTimeMs,
-          ...(classifiedState.classificationTimeMs != null && {
-            classificationTimeMs: classifiedState.classificationTimeMs,
-          }),
-          ...(finalState.searchTimeMs != null && { searchTimeMs: finalState.searchTimeMs }),
-        },
-      });
-
-      log.info(`[ChatGraph:Resume] Complete: ${fullText.length} chars in ${totalTimeMs}ms`);
-      sse.end();
-      return { status: 200 as const, body: undefined };
-    } catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-      const errorStack = error instanceof Error ? error.stack : undefined;
-      log.error(`[ChatGraph:Resume] Controller error: ${errorMessage}`);
-      if (errorStack) log.error(`[ChatGraph:Resume] Stack: ${errorStack}`);
-      if (!sse.isEnded()) {
-        sse.send('error', { error: PROGRESS_MESSAGES.internalError });
-        sse.end();
-      }
-      return { status: 200 as const, body: undefined };
-    }
+    return runChatGraphResume({ req: args.req, body: args.body, sse });
   },
 });
